Fix crash with Single<>
`Single<>` seems to assume that it will definitely be able to find some
entity that satisfies whatever is in `<>`.

When `ActivePiece` is removed from an entity, then we have no more
entities with the `ActivePiece` component, which causes any Systems that
have `Single<...,ActivePiece,...>` to crash.

Using `Query` we don't have that restriction, so the program doesn't
crash.

It is a bit ugly because of the required for loops even though I know
that there will ever only be 1 `ActivePiece`,
but I'm sure there's a better way for this that I can figure out later
on.