Fixes for drawing a piece and collision detection on bottom
`grid.rows - 1` because the rows go from 0..19 and `grid.rows == 20`. So
we were off-by-one before.

Also removed the `- block_size` because it made no sense. We should draw
the blocks exactly were they should be. This was causing issues with the
visualization of block position on the grid (e.g. grid_position.y was 0,
but the block was being drawn on row 1).